Escape special chars in search filter
This has been done in user_groups(), but not in authenticate().
Escaping special chars is necessary to prevent LDAP injection
attacks. In this case, the password check is not affected, but
this might still be abused in various ways.

Not that we don't use the auto_escape flag of ldap3, since it
still accepts some kinds of filter modification.